+++ com/eurekaAccounts/stocks/dao/LookUpStockFundamentals.java
@@ -33,7 +33,7 @@ public class LookUpStockFundamentals extends BaseDAO {
             stockFundementals1.setSectorId(resultSet.getBigDecimal("sector_id"));
             stockFundementals1.setTickerSymbol(resultSet.getString("ticker_symbol"));
             stockFundementals1.setSubSectorId(resultSet.getInt("subsector_id"));
-            stockFundementals1.setMarketCap(resultSet.getBigDecimal("market_cap"));
+            stockFundementals1.setMarketCap(resultSet.getLong("market_cap"));
             stockFundementals1.setCurrentRatio(resultSet.getFloat("current_ratio"));
             stockFundementals.add(stockFundementals1);
         }
